tests console.py

+++ tests/test_models/test_engine/test_file_storage.py
@@ -18,11 +18,10 @@ class test_fileStorage(unittest.TestCase):
             del storage._FileStorage__objects[key]
 
     def tearDown(self):
-        """ Remove storage file at end of tests """
-        try:
-            os.remove('file.json')
-        except:
-            pass
+        """Clean up conditions after each test"""
+        storage._FileStorage__objects = {}
+        if os.path.exists("file.json"):
+            os.remove("file.json")
 
     def test_obj_list_empty(self):
         """ __objects is initially empty """
